code: change type

+++ ConsoleApplication1/ConsoleApplication1.cpp
@@ -3,7 +3,7 @@
 int main()
 {
     std::cout << "Enter A and B: ";
-    int A, B;
+    double A, B;
     std::cin >> A >> B;
     std::cout << "A + B = " << A + B << '\n'
         << "A - B = " << A - B << '\n';
